♻️ refactor: update imports to use models from products and users modules
