+++ services/web/server/src/simcore_service_webserver/groups/_groups_rest.py
@@ -18,7 +18,7 @@ from servicelib.aiohttp.requests_validation import (
 
 from .._meta import API_VTAG
 from ..login.decorators import login_required
-from ..products import products_service, products_web
+from ..products import products_web
 from ..security.decorators import permission_required
 from ..utils_aiohttp import envelope_json_response
 from . import _groups_service
@@ -43,7 +43,7 @@ async def list_groups(request: web.Request):
     """
     List all groups (organizations, primary, everyone and products) I belong to
     """
-    product: products_service.Product = products_web.get_current_product(request)
+    product = products_web.get_current_product(request)
     req_ctx = GroupsRequestContext.model_validate(request)
 
     (
