+++ services/web/server/src/simcore_service_webserver/products/products_service.py
@@ -1,4 +1,3 @@
-from ._models import CreditResult, Product, ProductName
 from ._service import (
     get_credit_amount,
     get_product,
@@ -8,14 +7,8 @@ from ._service import (
     list_products,
     list_products_names,
 )
-from .errors import BelowMinimumPaymentError, ProductPriceNotDefinedError
 
 __all__: tuple[str, ...] = (
-    "BelowMinimumPaymentError",
-    "CreditResult",
-    "Product",
-    "ProductName",
-    "ProductPriceNotDefinedError",
     "get_credit_amount",
     "get_product",
     "get_product_stripe_info",
